+++ src/clean_air/base/exceptions.py
