--- src/clean_air/users/user_manager.py
@@ -1,10 +1,0 @@
-"""
-Class to handle management of Clean Air Project Users
-"""
-
-
-class UserManager():
-
-    def check_permissions(self, user_id):
-        # do something
-        pass
