--- src/clean_air/visualise/assets/data.py
@@ -1,92 +1,0 @@
-""" Aircraft track data files
-"""
-def get_coords1():
-
-    coords_1 = [[50.781174, -1.835518], [50.773556, -1.879044], [50.803822, -1.886494],
-                [50.836315, -1.8332669999999998], [50.865452000000005, -1.7733599999999998], [50.894978, -1.712547],
-                [50.925125, -1.650768], [50.95932, -1.5968069999999999], [50.997803000000005, -1.549045],
-                [51.035995, -1.500332], [51.074020000000004, -1.450753], [51.110039, -1.3976110000000002],
-                [51.144901000000004, -1.34208], [51.179660999999996, -1.286606], [51.214611, -1.229973],
-                [51.25029, -1.175453], [51.287357, -1.125261], [51.32505, -1.076755], [51.36359, -1.030339],
-                [51.40332, -0.9872790000000001], [51.443951, -0.948473], [51.485157, -0.913467],
-                [51.526333, -0.8779600000000001], [51.567261, -0.8407709999999999], [51.608482, -0.804758],
-                [51.621937, -0.735417], [51.628674, -0.663424], [51.635436999999996, -0.592378], [51.640484, -0.520518],
-                [51.644009000000004, -0.449144], [51.64185, -0.379973], [51.626282, -0.331458],
-                [51.63615, -0.34129899999999996], [51.633373, -0.291534], [51.65696, -0.27571],
-                [51.656609, -0.31852600000000003], [51.654716, -0.365444], [51.692669, -0.365267],
-                [51.705349, -0.305706], [51.697659, -0.236922], [51.689274, -0.167826], [51.682377, -0.098018],
-                [51.676762, -0.027268999999999998], [51.668499, 0.04258], [51.658302, 0.11105], [51.641857, 0.173259],
-                [51.611195, 0.215704], [51.580963, 0.260294], [51.548973, 0.301687], [51.515518, 0.32414299999999996],
-                [51.518425, 0.270473], [51.519913, 0.211358], [51.515361999999996, 0.152892],
-                [51.510459999999995, 0.09705599999999999], [51.501999, 0.04485], [51.493015, -0.006291],
-                [51.484108, -0.056930999999999995], [51.47612, -0.107525], [51.466621, -0.15825699999999998],
-                [51.456413, -0.20856599999999997], [51.447483, -0.259534], [51.437874, -0.311013],
-                [51.419865, -0.352619], [51.392371999999995, -0.378542], [51.366768, -0.400008],
-                [51.339127000000005, -0.424605], [51.319321, -0.386986], [51.324028000000006, -0.313579],
-                [51.326415999999995, -0.240614], [51.328518, -0.167846], [51.332684, -0.094951],
-                [51.341465, -0.022494999999999998], [51.368851, 0.029671], [51.354038, 0.050559],
-                [51.329575, 0.033283999999999994], [51.298607000000004, 0.017238999999999997],
-                [51.271328000000004, 0.048258999999999996], [51.300003000000004, 0.10741600000000001],
-                [51.337708, 0.154934], [51.374027000000005, 0.200801], [51.409908, 0.24485],
-                [51.444365999999995, 0.289588], [51.480942, 0.33534200000000003], [51.511204, 0.322129],
-                [51.513226, 0.26760100000000003], [51.538677, 0.251325], [51.537636, 0.321019], [51.514736, 0.337604],
-                [51.519363, 0.283482], [51.52882, 0.23352199999999998], [51.543453, 0.292518], [51.529434, 0.354299],
-                [51.518817999999996, 0.316446], [51.519905, 0.263967], [51.540436, 0.272861], [51.538292, 0.341898],
-                [51.520809, 0.32225], [51.518421000000004, 0.271337], [51.537071000000005, 0.24648499999999998],
-                [51.540276, 0.316577], [51.521415999999995, 0.335472], [51.519974, 0.279907], [51.517429, 0.223392],
-                [51.514606, 0.168559], [51.510696, 0.114373], [51.504642, 0.061107], [51.495971999999995, 0.008891],
-                [51.487202, -0.044238], [51.478103999999995, -0.09704199999999999], [51.469318, -0.14938800000000002],
-                [51.459796999999995, -0.201354], [51.451283000000004, -0.253921], [51.442078, -0.306958],
-                [51.432804, -0.36007100000000003], [51.423653, -0.413889], [51.415203000000005, -0.469074],
-                [51.405437, -0.521471], [51.393364, -0.5734560000000001], [51.381415999999994, -0.6253380000000001],
-                [51.372608, -0.674301], [51.365967, -0.7265199999999999], [51.360123, -0.782308],
-                [51.356441, -0.8404790000000001], [51.350933000000005, -0.89779], [51.342361, -0.9527030000000001],
-                [51.333527000000004, -1.007075], [51.319126000000004, -1.0574], [51.305065, -1.107566],
-                [51.294403, -1.160457], [51.284206, -1.21302], [51.273987, -1.266857], [51.263366999999995, -1.321526],
-                [51.243759000000004, -1.361587], [51.209034, -1.3759350000000001], [51.17889, -1.397691],
-                [51.150314, -1.425253], [51.122082, -1.453384], [51.088829, -1.449003], [51.116505, -1.454209],
-                [51.154167, -1.427834], [51.189098, -1.408911], [51.1754, -1.395684], [51.149421999999994, -1.425171],
-                [51.120045000000005, -1.4448379999999998], [51.088485999999996, -1.456016],
-                [51.057903, -1.4786290000000002], [51.028042, -1.507377], [50.99762, -1.536192], [50.966904, -1.564647],
-                [50.936073, -1.590018], [50.903656, -1.607407], [50.873184, -1.630891],
-                [50.843601, -1.6577400000000002], [50.811828999999996, -1.6766869999999998],
-                [50.80019, -1.7227720000000002], [50.792152, -1.771628], [50.785323999999996, -1.81017],
-                [50.780079, -1.8422020000000001], [50.780029, -1.8497310000000002]]
-    return coords_1
-
-    def get_coords_2():
-        coords_2 = [[50.780945, -1.8368659999999999], [50.780827, -1.8375700000000001], [50.780704, -1.838287],
-                [50.780582, -1.839015], [50.780457, -1.839751],
-                [50.780327, -1.840493], [50.780193, -1.8412400000000002], [50.780064, -1.8419889999999999],
-                [50.779934000000004, -1.84274], [50.779804, -1.843491],
-                [50.779675, -1.8442439999999998], [50.779540999999995, -1.8449959999999999],
-                [50.779408000000004, -1.8457459999999999], [50.779270000000004, -1.846495],
-                [50.779137, -1.8472419999999998], [50.778999, -1.8479869999999998], [50.778858, -1.84873],
-                [50.778721000000004, -1.849476], [50.778576, -1.85022],
-                [50.778431, -1.8509669999999998], [50.778286, -1.8517150000000002], [50.778141, -1.852468],
-                [50.777996, -1.8532240000000002],
-                [50.777846999999994, -1.853981], [50.777702000000005, -1.854734], [50.777557, -1.8554869999999999],
-                [50.777412, -1.856241],
-                [50.777266999999995, -1.8569950000000002], [50.777126, -1.857743], [50.776989, -1.8584889999999998],
-                [50.776852000000005, -1.8592349999999997],
-                [50.776714, -1.859979], [50.776577, -1.8607169999999997], [50.776443, -1.86145],
-                [50.776314, -1.8621779999999999], [50.77618, -1.862902], [50.776051, -1.863621],
-                [50.775925, -1.864336], [50.775803, -1.8650509999999998], [50.775688, -1.865762],
-                [50.775574, -1.866471], [50.775459000000005, -1.8671740000000001],
-                [50.775352000000005, -1.867873], [50.775242, -1.868568], [50.775135, -1.869262],
-                [50.775028000000006, -1.8699569999999999], [50.774921, -1.8706509999999998],
-                [50.774815000000004, -1.871346], [50.774704, -1.872039], [50.774592999999996, -1.872732],
-                [50.774479, -1.8734259999999998], [50.774361, -1.8741240000000001],
-                [50.774239, -1.8748209999999998], [50.774117, -1.87552], [50.773994, -1.87622],
-                [50.773872, -1.8769200000000001], [50.773754, -1.8776220000000001],
-                [50.773647, -1.8783290000000001], [50.773556, -1.879044], [50.773483, -1.8797689999999998],
-                [50.773434, -1.8805040000000002], [50.773410999999996, -1.88125],
-                [50.773415, -1.882005], [50.773449, -1.882768], [50.773514, -1.883536], [50.773609, -1.88431],
-                [50.773734999999995, -1.885088], [50.773895, -1.885864],
-                [50.774086, -1.886637], [50.774311, -1.8874060000000001], [50.774574, -1.888161],
-                [50.774871999999995, -1.8889], [50.775203999999995, -1.8896240000000002],
-                [50.775566, -1.8903279999999998], [50.775959, -1.8910080000000002], [50.776379, -1.891661],
-                [50.776825, -1.8922889999999999], [50.777294, -1.892899],
-                [50.777786, -1.8934819999999999], [50.778296999999995, -1.894035],
-                [50.778828000000004, -1.8945599999999998], [50.779381, -1.8950580000000001]]
-        return coords_2
