+++ tests/unit/base/test_models_duration.py
@@ -8,7 +8,7 @@ from datetime import datetime, timedelta
 import pytest
 from dateutil.relativedelta import relativedelta
 
-from clean_air.models import Duration
+from clean_air.base.models import Duration
 
 
 class DurationTest(unittest.TestCase):
